+++ landscape/utils.py
@@ -11,22 +11,9 @@ def namespace_exists(namespace):
         return False
 	
 
-def kubernetes_get_context():
-    k8s_get_context_cmd = "kubectl config current-context"
-    proc = subprocess.Popen(k8s_get_context_cmd, stdout=subprocess.PIPE, shell=True)
-    k8s_context = proc.stdout.read().rstrip()
-    return k8s_context
-
-
 def git_get_branch():
 	return subprocess.check_output(['git', 'symbolic-ref', 'HEAD', '--short' ]).rstrip()
 
-
-def kubernetes_set_context(k8s_context):
-    if subprocess.call(['kubectl', 'config', 'use-context', k8s_context], shell=True):
-        return True
-    else:
-        return False
 
 def test_dns_domain(k8s_provisioner, cluster_dns_domain):
     """
